new text renderer, various changes
	 * New changes to text rendering - it now supports inline color,
size,and formats! Text boxes are broken as of this moment - will be
fixed soon. It will only be pushed to master when text boxes are fixed
	*Texture binding - similarly to the addition of render modes, the
JumboTextureBinder will now handle any texture binding needs! This was
added for multiple reasons:
	 1. This allows for someone to create their own texture system, and not
use the default OpenGL system
	 2. Someone might want to use a different method of storing textures
with OpenGL
	 3. To create a no-graphics mode, where you can use non-graphical
classes without it raising a runtimeerror (because of no opengl context)
	 
Also, as of this point, all opengl code in the engine can be overriden
by the user without having to change the source code. This opens a new
level of versatility to the engine, as you can now use any graphics
system that you wish! DirectX, Vulkan, JavaFX are now possible with the
engine (not be built-in)
	*Now all entities have an active variable. By default, all non-input
classes will determine whether to use their custom tick action based on
whether they are active. Also, animations will 'pause' if they are
inactive
	*Various bug fixes
	*JumboColor - replaces the previous built-in Java color class, for
extra control and synergy with Opengl (like converting hexadecimal
codes, and creating Opengl-accepted color bytes)
	*Exceptions - now classes will throw exceptions, like the new
JumboException, if they are not initiations
	*Console logging - JumboConsole has been improved to also provide a
logging system. Custom commands are still a WIP
	*Custom error messages - when calling JumboErrorHandler.handle, you can
now define a custom error message. The only thing this does currently is
print it to console and put it in the log. I am planning to add a small
window that pops up with the error message for the user to see.