removed rectangle+wireframe from settings, gl version
* Rectangle and wireframe modes have been removed from JumboSettings.
Their purpose can be accomplished through JumboRenderModes, which is
better.
* The OpenGL version before wouldn't allow anything other than 1.1 . Now
it can support all OpenGL versions so you can mix and match
JumboRenderModes
* Fixed a check in addRenderMode(mode, index) in JumboRenderer which
would improperly raise an error. The error would be raised every time
this is called, as it would check for index being smaller than buffer
size. Index should be smaller than size, and thus it was fixed to only
raise errors when index is larger than size.